backend/requirements.txt
aiofiles==23.2.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==3.2.2
billiard==4.2.4
black==23.11.0
boto3==1.29.7
botocore==1.32.7
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
coverage==7.13.1
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
filelock==3.20.3
flake8==6.1.0
fonttools==4.61.1
fsspec==2026.1.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.2
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mpmath==1.3.0
mypy==1.7.0
mypy_extensions==1.1.0
networkx==3.6.1
ninja==1.13.0
numpy==1.26.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.92
packaging==25.0
pandas==2.1.3
passlib==1.7.4
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
polars==1.37.1
polars-runtime-32==1.37.1
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyclipper==1.4.0
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyflakes==3.1.0
pyparsing==3.3.1
pytesseract==0.3.13
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.7
requests==2.32.5
rsa==4.9.1
s3transfer==0.8.0
scikit-image==0.26.0
scikit-learn==1.3.2
scipy==1.16.3
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.9
ultralytics-thop==2.0.18
urllib3==2.0.7
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
